+++ lib/src/quantity/src/number/roman_numerals.dart
@@ -1,4 +1,20 @@
+/// A class to convert between Roman and decimal numerals.
+///
+/// The class provides functionality to represent an integer as a Roman numeral
+/// and vice versa. For numbers greater than 3999, a bracket notation is used
+/// to represent multiplication by 1000.
+///
+/// Example:
+///
+/// ```dart
+/// var numeral = RomanNumerals(4567);
+/// print(numeral); // Outputs: (IV)DLXVII
+///
+/// var numeralFromRoman = RomanNumerals.fromRoman('(IV)DLXVII');
+/// print(numeralFromRoman.value); // Outputs: 4567
+/// ```
 class RomanNumerals {
+  /// A private constant map that represents the standard numerals.
   static const Map<String, int> _numerals = {
     'M': 1000,
     'CM': 900,
@@ -15,35 +31,37 @@ class RomanNumerals {
     'I': 1
   };
 
-  /// Extended Roman numerals
-  static const Map<String, int> _extendedNumerals = {
-    '_M': 1000000, // Represents an overlined M (1,000,000)
-    '_D': 500000, // Represents an overlined D (500,000)
-    '_C': 100000, // Represents an overlined C (100,000)
-    '_L': 50000, // Represents an overlined L (50,000)
-    '_X': 10000, // Represents an overlined X (10,000)
-    '_V': 5000, // Represents an overlined V (5,000)
-  };
-
-  // Combining standard and extended numerals
-  static final Map<String, int> _allNumerals = {
-    ..._extendedNumerals,
-    ..._numerals,
-  };
-
+  /// The integer value represented by this instance.
   final int value;
 
+  /// Private cache to improve performance for the conversion methods.
   static final Map<int, String> _numToRomanCache = {};
   static final Map<String, int> _romanToNumCache = {};
 
+  /// Constructs a [RomanNumerals] object from a given integer [value].
+  ///
+  /// Throws an [ArgumentError] if [value] is negative.
+  ///
+  /// Example:
+  /// ```dart
+  /// print(RomanNumerals(69)); // LXIX
+  /// print(RomanNumerals(8785)); // (VIII)DCCLXXXV
+  /// ```
   RomanNumerals(this.value) {
     if (value < 0) {
       throw ArgumentError('Roman numerals cannot represent negative numbers.');
     }
   }
 
+  /// Constructs a [RomanNumerals] object from a given Roman numeral [roman].
+  ///
+  /// Example:
+  /// ```dart
+  /// var numeral = RomanNumerals.fromRoman('XIV'); // 14
+  /// ```
   RomanNumerals.fromRoman(String roman) : value = _fromRoman(roman);
 
+  /// A private method to convert a Roman numeral string to an integer.
   static int _fromRoman(String roman) {
     roman = roman.toUpperCase();
 
@@ -54,24 +72,37 @@ class RomanNumerals {
       int i = roman.length - 1;
 
       while (i >= 0) {
-        String currentSymbol =
-            (i > 0 && _allNumerals.containsKey(roman.substring(i - 1, i + 1)))
-                ? roman.substring(i - 1, i + 1)
-                : roman[i];
+        // Check if the current character is a parenthesis
+        if (roman[i] == ')') {
+          // Find the matching opening parenthesis
+          int j = roman.lastIndexOf('(', i);
+          if (j == -1) {
+            throw ArgumentError('Invalid Roman numeral: $roman');
+          }
 
-        int? currentValue = _allNumerals[currentSymbol];
-        if (currentValue == null) {
-          throw ArgumentError('Invalid Roman numeral: $roman');
-        }
+          // Convert the part of the string inside the parentheses
+          String innerRoman = roman.substring(j + 1, i);
+          int innerValue = _fromRomanWithoutParentheses(innerRoman);
+          result += innerValue * 1000;
 
-        if (currentValue < prevValue) {
-          result -= currentValue;
+          i = j - 1;
         } else {
-          result += currentValue;
-        }
-        prevValue = currentValue;
+          String currentSymbol = roman[i];
 
-        i -= currentSymbol.length;
+          int? currentValue = _numerals[currentSymbol];
+          if (currentValue == null) {
+            throw ArgumentError('Invalid Roman numeral: $roman');
+          }
+
+          if (currentValue < prevValue) {
+            result -= currentValue;
+          } else {
+            result += currentValue;
+          }
+          prevValue = currentValue;
+
+          i--;
+        }
       }
 
       // Cache the converse for improved performance
@@ -81,7 +112,41 @@ class RomanNumerals {
     });
   }
 
-  /// Converts an integer to a Roman numeral string.
+  /// A helper private method to convert a Roman numeral string
+  /// (without parentheses) to an integer.
+  static int _fromRomanWithoutParentheses(String roman) {
+    int result = 0;
+    int prevValue = 0;
+    int i = roman.length - 1;
+
+    while (i >= 0) {
+      String currentSymbol = roman[i];
+
+      int? currentValue = _numerals[currentSymbol];
+      if (currentValue == null) {
+        throw ArgumentError('Invalid Roman numeral: $roman');
+      }
+
+      if (currentValue < prevValue) {
+        result -= currentValue;
+      } else {
+        result += currentValue;
+      }
+      prevValue = currentValue;
+
+      i--;
+    }
+
+    return result;
+  }
+
+  /// Converts the integer value of this instance to a Roman numeral string.
+  ///
+  /// Example:
+  /// ```dart
+  /// var numeral = RomanNumerals(14);
+  /// print(numeral.toRoman());  // Outputs: XIV
+  /// ```
   String toRoman() {
     // Using cache's putIfAbsent to streamline caching
     return _numToRomanCache.putIfAbsent(value, () {
@@ -93,13 +158,15 @@ class RomanNumerals {
       int number = value;
       StringBuffer result = StringBuffer();
 
-      // Loop through the combined numerals, from largest to smallest
-      for (var entry in _allNumerals.entries) {
-        while (number >= entry.value) {
-          number -= entry.value;
-          result.write(entry.key);
-        }
+      // If the value is more than 3999, use brackets for the thousands part
+      if (value > 3999) {
+        int thousands = value ~/ 1000;
+        number %= 1000;
+        result.write('(${_toRomanWithoutParentheses(thousands)})');
       }
+
+      // Then handle the rest of the number
+      result.write(_toRomanWithoutParentheses(number));
 
       // Cache the converse for improved performance
       _romanToNumCache[result.toString()] = value;
@@ -108,7 +175,32 @@ class RomanNumerals {
     });
   }
 
-  /// Validates if the string is a valid Roman numeral.
+  /// A helper private method to convert an integer to a Roman numeral string
+  /// without using parentheses.
+  String _toRomanWithoutParentheses(int value) {
+    int number = value;
+    StringBuffer result = StringBuffer();
+
+    // Loop through the numerals, from largest to smallest
+    for (var entry in _numerals.entries) {
+      while (number >= entry.value) {
+        number -= entry.value;
+        result.write(entry.key);
+      }
+    }
+
+    return result.toString();
+  }
+
+  /// Validates if a given string [roman] is a valid Roman numeral.
+  ///
+  /// Returns `true` if [roman] is a valid Roman numeral, `false` otherwise.
+  ///
+  /// Example:
+  /// ```dart
+  /// print(RomanNumerals.isValidRoman('XIV'));  // Outputs: true
+  /// print(RomanNumerals.isValidRoman('XIVV')); // Outputs: false
+  /// ```
   static bool isValidRoman(String roman) {
     try {
       _fromRoman(roman);
@@ -118,19 +210,26 @@ class RomanNumerals {
     }
   }
 
-  // Operator overrides
+  /// Adds the values of two [RomanNumerals] objects and returns a new one.
   RomanNumerals operator +(RomanNumerals other) {
     return RomanNumerals(value + other.value);
   }
 
+  /// Subtracts the value of another [RomanNumerals] from this one and returns a new one.
   RomanNumerals operator -(RomanNumerals other) {
+    if (value - other.value < 0) {
+      throw ArgumentError(
+          'Result of subtraction cannot be a negative value in Roman numerals.');
+    }
     return RomanNumerals(value - other.value);
   }
 
+  /// Multiplies the values of two [RomanNumerals] objects and returns a new one.
   RomanNumerals operator *(RomanNumerals other) {
     return RomanNumerals(value * other.value);
   }
 
+  /// Divides the value of this [RomanNumerals] by another and returns a new one.
   RomanNumerals operator /(RomanNumerals other) {
     if (other.value == 0) {
       throw ArgumentError('Division by zero is not allowed.');
@@ -178,10 +277,6 @@ class RomanNumerals {
     return RomanNumerals(value >> shift);
   }
 
-  RomanNumerals operator ~() {
-    return RomanNumerals(~value);
-  }
-
   @override
   bool operator ==(Object other) {
     if (other is RomanNumerals) {
@@ -193,6 +288,7 @@ class RomanNumerals {
   @override
   int get hashCode => value.hashCode;
 
+  /// Returns the Roman numeral string representation of this instance.
   @override
   String toString() {
     return toRoman();